Split off menu_ui module
