+++ iris_core/modules/extra/menu_ui/menu_ui.iris.module
@@ -1,0 +1,3 @@
+{
+  "name": "Menu UI"
+}
